move component tests into their own files
